+++ package.json
@@ -6,7 +6,8 @@
     "build": "webpack --config webpack.config.js",
     "build:watch": "webpack --watch --config webpack.config.js",
     "start-server": "nodemon --inspect server/index.js",
-    "test": "jest ./spec"
+    "test": "jest",
+    "test:coverage": "jest --coverage"
   },
   "repository": {
     "type": "git",
@@ -28,8 +29,11 @@
     "eslint-plugin-react": "^7.30.1",
     "eslint-plugin-react-hooks": "^4.6.0",
     "jest": "^28.1.2",
+    "jest-config": "^28.1.3",
+    "jest-environment-jsdom": "^28.1.3",
     "mocha": "^10.0.0",
     "nodemon": "2.0.19",
+    "react-test-renderer": "^18.2.0",
     "webpack": "^5.73.0",
     "webpack-cli": "^4.10.0"
   },
